update: mv order_settings to setting_service
